# ranking: Improve export query performance